refactor(LocalAddressBook): move toolbar handlers into SectionToolbar
