+++ src/pages/options/AddressesSettings/SectionToolbar.tsx
@@ -1,4 +1,4 @@
-import { ChangeEvent, MouseEvent } from 'react';
+import { ChangeEvent, MouseEvent, useCallback } from 'react';
 import { Box, Button, TextField } from '@mui/material';
 import {
   ContentPaste as ContentPasteIcon,
@@ -6,49 +6,112 @@ import {
   Save as SaveIcon,
   Download as DownloadIcon,
 } from '@mui/icons-material';
+import * as murmurhash from 'murmurhash';
 
-interface AddressBookToolbarProps {
+interface SectionToolbarProps {
   fetchUrl: string;
-  onUrlChange: (e: ChangeEvent<HTMLInputElement>) => void;
-  onFetch: (e: MouseEvent) => void;
-  onPaste: (e: MouseEvent) => void;
-  onRevert: (e: MouseEvent) => void;
-  onSave: (e: MouseEvent) => void;
+  setFetchUrl: (url: string) => void;
+  updateUrl: (url: string) => void;
+  setLabels: (labels: string) => void;
+  setCurrentLabelsHash: (hash: number) => void;
+  validate: (content: string) => void;
+  handleSave: () => void;
+  getSection: () => void;
   canRevert: boolean;
   canSave: boolean;
 }
 
-export default function AddressBookToolbar({
+export default function SectionToolbar({
   fetchUrl,
-  onUrlChange,
-  onFetch,
-  onPaste,
-  onRevert,
-  onSave,
+  setFetchUrl,
+  updateUrl,
+  setLabels,
+  setCurrentLabelsHash,
+  validate,
+  handleSave,
+  getSection,
   canRevert,
   canSave,
-}: AddressBookToolbarProps) {
+}: SectionToolbarProps) {
+  const handleUrlChange = useCallback(
+    (e: ChangeEvent<HTMLInputElement>) => {
+      const newUrl = e.target.value;
+      setFetchUrl(newUrl);
+      updateUrl(newUrl);
+    },
+    [setFetchUrl, updateUrl],
+  );
+
+  const handleFetch = useCallback(
+    async (e: MouseEvent) => {
+      e.stopPropagation();
+      try {
+        const response = await fetch(fetchUrl);
+        if (!response.ok) {
+          throw new Error(`HTTP error! status: ${response.status}`);
+        }
+        const content = await response.text();
+        setLabels(content);
+        const hash = murmurhash.v3(content);
+        setCurrentLabelsHash(hash);
+        validate(content);
+      } catch (error) {
+        console.error('Error fetching URL:', error);
+        validate('Error fetching URL: ' + (error instanceof Error ? error.message : String(error)));
+      }
+    },
+    [fetchUrl, setLabels, setCurrentLabelsHash, validate],
+  );
+
+  const handlePaste = useCallback(
+    async (e: MouseEvent) => {
+      e.stopPropagation();
+      const clipboardContents = await window.navigator.clipboard.readText();
+      const hash = murmurhash.v3(clipboardContents);
+      setLabels(clipboardContents);
+      setCurrentLabelsHash(hash);
+      validate(clipboardContents);
+    },
+    [setLabels, setCurrentLabelsHash, validate],
+  );
+
+  const handleRevert = useCallback(
+    (e: MouseEvent) => {
+      e.stopPropagation();
+      getSection();
+    },
+    [getSection],
+  );
+
+  const handleSaveClick = useCallback(
+    (e: MouseEvent) => {
+      e.stopPropagation();
+      handleSave();
+    },
+    [handleSave],
+  );
+
   return (
     <Box sx={{ display: 'flex', alignItems: 'center', gap: 1, mb: 1 }}>
       <TextField
         size="small"
         placeholder="Enter URL to fetch addresses"
         value={fetchUrl}
-        onChange={onUrlChange}
+        onChange={handleUrlChange}
         sx={{ flexGrow: 1 }}
       />
       <Button
         variant="contained"
         startIcon={<DownloadIcon />}
         size="small"
-        onClick={onFetch}
+        onClick={handleFetch}
         disabled={!fetchUrl.trim()}>
         Fetch
       </Button>
       <Button
         className="section-paste-button"
         variant="contained"
-        onClick={onPaste}
+        onClick={handlePaste}
         startIcon={<ContentPasteIcon />}
         size="small"
         sx={{ mr: 1 }}>
@@ -57,7 +120,7 @@ export default function AddressBookToolbar({
       <Button
         className="section-revert-button"
         variant="contained"
-        onClick={onRevert}
+        onClick={handleRevert}
         startIcon={<RestorePageIcon />}
         disabled={!canRevert}
         size="small"
@@ -67,7 +130,7 @@ export default function AddressBookToolbar({
       <Button
         className="section-save-button"
         variant="contained"
-        onClick={onSave}
+        onClick={handleSaveClick}
         startIcon={<SaveIcon />}
         disabled={!canSave}
         size="small"
